terminal session shortcuts badly broken
rename does not work via shortcut (F2) nor abbreviation (tren), so
removed them to not give the illusion they work.

tab switching via CTRL+circumflex "works" but drops a literal circumflex
to the terminal through to the terminal that was active when using the
shortcut. So better don't use this here (it can't be re-assigned
explicitly for this as this shortcut is valid for all tool windows with
tabs)

tab switching via ALT+down works though.